[es5] Parsing for es5\JS based on original JS parsing.
Start from the base \JS parser.  Objects are different and have a
notion of attributes.

Add GetFieldSurface and UpdateFieldSurface to the syntax of es5\JS.
These are useful because we don't necessarily have references to the
objects as we are parsing.  We can use these forms to translate into
the normal forms that store a this reference in the rule.