[es5] Start of interpreter without delta function.  Changes to objects
es5_eval has the interpreter without a delta function.

Evaluation values can be references to hash structures representing
objects, references (representing variables) of values, closures, and
the constants defined in JavaScript_syntax.

VarCells aren't really values in that they are only produced for let,
fix, and function application.  EId and ESet look up VarCells in the
environment and dereference or set their contents respectively.

EUpdateField and EDeleteField change the value inside the reference in
an ObjCell.  EGetField, EUpdateField, and EDeleteField are the only
rules that can currently look inside an ObjCell.

ERef is gone, instead, object literals are no longer values.  Object
literal syntax has a rule in the reduction relation that adds an
object value to the heap, and puts the location in the evaluation
context as the resulting value.