[es5] More additions to the delta function and to the environment.
Notably, get_own_property is implemented, to get back an object with
properties representing the attributes (getter, setter, config, etc.)
of a particular property.

More primops, like typeof and surface-typeof, work and can be used in
tests.

Changed a bunch of identifiers from Object.prototype ->
Object_prototype (and similar for function and array), so it's clear
that it's a variable and not a property access each time.  Todo is to
differentiate the cases where these need to be accessed via getters.
For example, new Object() says the prototype is set to "the standard
built-in Object prototype object", *not* to do Object["prototype"],
while the default [[construct]] behavior is to actually use [[Get]] to
get the prototype property.