[es5] Operations, primitives, and moved arguments back to desugaring.
In es5-lib, implemented:

- instanceof
- ToNumber
- ToString
- [[plus]]
- Object.getPrototypeOf
- TypeError
- Function.prototype.apply
- Function.prototype.call

In es5_delta, added get_proto, which gets proto properties out of
objects.  This is in the spec as Object.getPrototypeOf.

Moved the arguments object out of the semantics (es5_eval).  It was
much cleaner than dealing with the implicit creation of the arguments
object when doing things like [[apply]] and [[call]].  Also, having
the arguments object in the semantics makes the *semantics* dependent
on the existence of Object.prototype, which seems like an unacceptable
abstraction breakage.

Cleaned up the tests in tests/eval_es5, and wrote a script to run them
all at once.  These are *really dumb* tests, and should be used to
sanity check changes in eval and desugaring to make sure something
silly didn't break.