[es5] Change to arguments object model.
Now, *all* arguments objects are made via desugaring.  This makes it
so the interpreter doesn't have to know anything about
Object_prototype or other values in the environment (the way it should
be).

To do this for getters and setters, which were being handled before
with arguments objects created by the semantics, we desugar field
lookup and field update to create an arguments object that will
*potentially* be used if a getter or setter is encountered.

Creating this object has no observable side effects.  It does have the
unobservable side effect of creating a new heap reference, but no
references to the created arguments object will be exposed unless a
getter or setter is called.  If this were really a problem, it could
be a thunk that creates the object instead.

I'm not completely thrilled with this, since it adds another layer of
complexity to EGetField and ESetField, but I much prefer it to pushing
environment information into the semantics.

Also, took a number of common methods out of es5_desugar and put them
in desugar_helpers, which the parser uses now as well.