[es5] More changes to args objects.
In getters and setters, we don't know the callee until we actually
find the getter or setter itself.  So, store a function in the
arguments position that takes the function and sets it to be the
arguments objects' callee field when it is found.

I also took out a rule from ds_global for applications, and fixed the
case where a this-value isn't provided to an application to correctly
give [[global]] as this.  Some of test cases were dependent on this
funky behavior, and I updated them accordingly.