[es5] Add terms to the language for getting attributes directly.
Added two new types of terms, here are examples in \JS syntax:

1. Attribute lookup:

   obj["foo"<:writable>]
   obj["bar"<:configurable>]

2. Attribute setting:

   obj["foo"<:writable> = true]
   obj["bar"<:configurable> = false]

The goal is to have them do, much more directly, what was being
simulated with objects in defineOwnProperty and
getOwnPropertyDescriptor.

They will maintain the monotinicity and constraints imposed by
[[configurable]], [[writable]], and [[extensible]].  This doesn't
matter for attribute lookup, but it does for attribute setting.  In
particular, whenever you update a property's attributes, you need to
check some combination of config, writable, and extensible.  Making
this explicit in the semantics will make it easier to obviously show
that the guarantees intended by freeze, seal, preventExtensions, and
similar are being enforced.