[es5] Add attribute access and setting to the interpeter and es5-lib
Got rid of two primitives: get-own-property and define-property, in
favor of adding set_attr and get_attr to the interpreter.

In es5-lib, references to these can be removed, and can directly set
the properties to the relevant values.

The rule for property setting in the interpreter makes sure that you
don't go having a setter and a writable, or a getter and a value, on
the same property.  Object.defineOwnProperty in the environment checks
for this, too, so it can throw the correct error.

However, we get a guarantee in the semantics that this sort of wonky
behavior can't happen.  defineOP and getOwnPD are the environment
interfaces for these parts of the semantics.