[es5] Array length getter and setter.
When the Array() constructor is called, the returned object gets a
pair of functions as a setter and getter for length.

These (mostly) mimic the behavior in the Array chapter of the spec,
without having to define crazy new [[defineProperty]] internals for
Arrays.

Will have to edit defineProperty, freeze, seal, and getOwnPD to lie
about the attributes of the length property (which is supposed to look
like a data property).

In the spec, if length isn't #writable, you aren't allowed to add
elements at greater indices than the value of length.  The method
implemented here doesn't disallow that.