[es5] Parsing, exprjs, and desugar support for getters and setters.
Added syntactic/desugaring support for getters and setters in object
literals.

Since we also want to be able to use get and set as property names, I
added specific cases for using "get" and "set" as PropStrings.  There
might be a better way to do that, but this seems to work fine.

Getters and setters must have zero and one arguments respectively ---
this is what the productions in 11.1.5 require, though rhino and
browsers don't seem hindered by this restriction.

Converting to exprjs is boring, but desugaring for object literals
actually has to change in a non-trivial way.  Since there can be more
than one *attribute* per property inside the object literal, an extra
desugaring step collects the attributes per property.

Some extra work is needed to screen out the "bad" cases of including
multiple getters/setters for the same property.  But this can't go too
wrong, because the defaults for property attributes are to be
configurable and enumerable by default.