[es5] Added a special Fail value for tracking certain errors
There wasn't a good way to handle the strict mode requirements on
throwing TypeErrors---it isn't a good idea to randomly have something
from the environment inside the semantics, but it also needs to be the
case that things like setting fields can be desugared into throwing
the right error in the event of unwritable properties and the like.

So there's a value called Fail that's a special error, and we can
desugar into things like:

let x = o["foo" = v]
    if(prim("fail?", x)) ThrowTypeError();
    else x

This is an alternative to something like:

if o["foo", #writable]
   o["foo" = v]
else
   ThrowTypeError();

Because we want the semantics to handle some of these issues, report
errors, and then do the right thing at the JavaScript level.