For-in loops
The meat of the loop is implemented in the environment as [[forin]].

The expression

for(x in obj) {
      body
}

desugars to

forin(desugar obj, (lambda (y) (set! x y)), (lambda () desugar body))

[[forin]] gets the names of the properties of obj and iterates over
them, checking to make sure they still exist and are still enumerable.

This jives with the semantics --- it requires that you not enumerate
fields that may get deleted, but you don't have to keep track of
fields that are added, so starting with an initial list is fine.

Both the primitive "property-names" and the forin function check to
make sure the property is #enumerable before doing anything with it.