ES5Conform, and scripts to run the tests.
We pass something like 11% of the tests in TestCases.

LJSElided includes a number of tests that involve eval and parsing
that we're uninterested in.  Perhaps we'll worry about these later,
but for now, semicolon insertion and eval are not major concerns.

This script:

/tests/ES5Conform/SimpleTestHarness/strict-tests.sh

Runs all the tests in the TestCases/ directory.  It relies on
ljs-test-single.sh, ljs-run.js, and ljs-harness.js.  This conformance
suite was originally designed to be run from a browser --- these
js files and scripts run the tests offline.